requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
aiohttp=3.8.1=py39h2bbff1b_1
aioredis=1.3.1=pypi_0
aiosignal=1.2.0=pyhd3eb1b0_0
amqp=5.0.9=pypi_0
argon2-cffi=21.3.0=pyhd8ed1ab_0
argon2-cffi-bindings=21.2.0=py39hb82d6ee_2
asgi-redis=1.1.0=pypi_0
asgiref=3.4.1=pyhd3eb1b0_0
asttokens=2.0.5=pyhd3eb1b0_0
async-timeout=4.0.1=pyhd3eb1b0_0
attrs=21.2.0=pypi_0
autobahn=21.11.1=pypi_0
automat=20.2.0=pypi_0
backcall=0.2.0=pyhd3eb1b0_0
backports-zoneinfo=0.2.1=pypi_0
beautifulsoup4=4.11.1=pyha770c72_0
billiard=3.6.4.0=py39h2bbff1b_0
black=22.3.0=pypi_0
blas=1.0=mkl
bleach=5.0.0=pyhd8ed1ab_0
blinker=1.4=py39haa95532_0
boto3=1.20.26=pypi_0
botocore=1.23.26=pypi_0
bottleneck=1.3.4=py39h080aedc_0
brotli=1.0.9=h8ffe710_7
brotli-bin=1.0.9=h8ffe710_7
brotlipy=0.7.0=py39h2bbff1b_1003
ca-certificates=2021.10.8=h5b45459_0
celery=5.1.2=pypi_0
certifi=2021.10.8=py39hcbf5309_2
cffi=1.15.0=py39h2bbff1b_1
channels=3.0.4=pypi_0
channels-redis=3.3.1=pypi_0
charset-normalizer=2.0.9=pypi_0
click=7.1.2=pypi_0
click-didyoumean=0.0.3=pyhd3eb1b0_0
click-plugins=1.1.1=pyhd3eb1b0_0
click-repl=0.2.0=pyhd3eb1b0_0
colorama=0.4.4=pyhd3eb1b0_0
conda=4.12.0=py39hcbf5309_0
conda-package-handling=1.8.1=py39h8cc25b3_0
console_shortcut=0.1.1=4
constantly=15.1.0=pypi_0
cpuonly=2.0=0
cryptography=36.0.1=pypi_0
cssselect=1.1.0=pypi_0
cycler=0.11.0=pyhd8ed1ab_0
daphne=3.0.2=pypi_0
debugpy=1.5.1=py39hd77b12b_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd8ed1ab_0
django=3.2=pypi_0
django-ckeditor=6.2.0=pypi_0
django-colorfield=0.6.0=pypi_0
django-js-asset=1.2.2=pypi_0
django-storages=1.12.3=pypi_0
django-timezone-field=4.2.1=pypi_0
django-tz-detect=0.4.0=pypi_0
djangorestframework=3.13.1=pyhd8ed1ab_0
entrypoints=0.4=py39haa95532_0
executing=0.8.3=pyhd3eb1b0_0
flit-core=3.7.1=pyhd8ed1ab_0
fonttools=4.33.3=py39hb82d6ee_0
freetype=2.10.4=h546665d_1
frozenlist=1.2.0=py39h2bbff1b_0
hiredis=2.0.0=pypi_0
hyperlink=21.0.0=pypi_0
icc_rt=2019.0.0=h0cc432a_1
icu=68.2=h0e60522_0
idna=3.3=pyhd3eb1b0_0
importlib-metadata=4.11.3=py39haa95532_0
importlib_resources=5.7.1=pyhd8ed1ab_0
incremental=21.3.0=pypi_0
intel-openmp=2021.4.0=haa95532_3556
ipykernel=6.9.1=py39haa95532_0
ipympl=0.9.1=pyhd8ed1ab_0
ipython=8.3.0=py39haa95532_0
ipython_genutils=0.2.0=py_1
ipywidgets=7.7.0=pyhd8ed1ab_0
isort=5.10.1=pypi_0
jedi=0.18.1=py39haa95532_1
jinja2=3.0.3=pyhd3eb1b0_0
jmespath=0.10.0=pypi_0
joblib=1.1.0=pyhd3eb1b0_0
jpeg=9e=h8ffe710_1
jsonschema=4.5.1=pyhd8ed1ab_0
jupyter_client=7.2.2=py39haa95532_0
jupyter_core=4.10.0=py39haa95532_0
jupyterlab_pygments=0.2.2=pyhd8ed1ab_0
jupyterlab_widgets=1.1.0=pyhd8ed1ab_0
kiwisolver=1.4.2=py39h2e07f2f_1
kombu=5.2.2=pypi_0
krb5=1.19.2=h5b6d351_0
libbrotlicommon=1.0.9=h8ffe710_7
libbrotlidec=1.0.9=h8ffe710_7
libbrotlienc=1.0.9=h8ffe710_7
libclang=11.1.0=default_h5c34c98_1
libpng=1.6.37=h1d00b33_2
libpq=12.9=hb652d5d_1
libtiff=4.0.10=h21b02b4_1005
libuv=1.40.0=he774522_0
libwebp=1.2.2=h57928b3_0
libwebp-base=1.2.2=h8ffe710_1
lxml=4.7.1=pypi_0
lz4-c=1.9.3=h8ffe710_1
mailchimp-marketing=3.0.75=pypi_0
markupsafe=2.0.1=py39h2bbff1b_0
matplotlib=3.5.2=py39hcbf5309_0
matplotlib-base=3.5.2=py39h581301d_0
matplotlib-inline=0.1.2=pyhd3eb1b0_2
menuinst=1.4.18=py39h59b6b97_0
mistune=0.8.4=py39hb82d6ee_1005
mkl=2021.4.0=haa95532_640
mkl-service=2.4.0=py39h2bbff1b_0
mkl_fft=1.3.1=py39h277e83a_0
mkl_random=1.2.2=py39hf11a4ad_0
msgpack=1.0.3=pypi_0
msgpack-python=0.5.6=pypi_0
multidict=5.2.0=pypi_0
munkres=1.1.4=pyh9f0ad1d_0
mypy-extensions=0.4.3=pypi_0
nbclient=0.6.3=pyhd8ed1ab_0
nbconvert=6.5.0=pyhd8ed1ab_0
nbconvert-core=6.5.0=pyhd8ed1ab_0
nbconvert-pandoc=6.5.0=pyhd8ed1ab_0
nbformat=5.4.0=pyhd8ed1ab_0
nest-asyncio=1.5.5=py39haa95532_0
networkx=2.7.1=pyhd3eb1b0_0
nltk=3.7=pypi_0
notebook=6.4.11=pyha770c72_0
numexpr=2.8.1=py39hb80d3ca_0
numpy=1.21.5=py39h7a0a035_2
numpy-base=1.21.5=py39hca35cd5_2
oauthlib=3.1.1=pypi_0
openssl=1.1.1o=h8ffe710_0
packaging=21.3=pyhd8ed1ab_0
pandas=1.4.2=py39hd77b12b_0
pandoc=2.18=h57928b3_0
pandocfilters=1.5.0=pyhd8ed1ab_0
parso=0.8.3=pyhd3eb1b0_0
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.4.0=pypi_0
pip=21.2.4=py39haa95532_0
platformdirs=2.5.2=pypi_0
powershell_shortcut=0.0.1=3
prometheus_client=0.14.1=pyhd8ed1ab_0
prompt-toolkit=3.0.24=pypi_0
prompt_toolkit=3.0.20=hd3eb1b0_0
psycopg2=2.9.2=pypi_0
pure_eval=0.2.2=pyhd3eb1b0_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycosat=0.6.3=py39h2bbff1b_0
pycparser=2.21=pyhd3eb1b0_0
pyg=2.0.4=py39_torch_1.11.0_cpu
pygad=2.18.1=pypi_0
pygments=2.11.2=pyhd3eb1b0_0
pyjwt=2.1.0=py39haa95532_0
pyopenssl=21.0.0=pyhd3eb1b0_1
pyparsing=3.0.9=pyhd8ed1ab_0
pyqt=5.12.3=py39hb0d2dfa_4
pyqt5-sip=4.19.18=pypi_0
pyqtchart=5.12=pypi_0
pyqtwebengine=5.12.1=pypi_0
pyquery=1.4.3=pypi_0
pyrsistent=0.18.1=py39hb82d6ee_1
pysocks=1.7.1=py39haa95532_0
python=3.9.12=h6244533_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-fastjsonschema=2.15.3=pyhd8ed1ab_0
python-louvain=0.15=pyhd3eb1b0_0
python_abi=3.9=2_cp39
pytorch=1.11.0=py3.9_cpu_0
pytorch-cluster=1.6.0=py39_torch_1.11.0_cpu
pytorch-mutex=1.0=cpu
pytorch-scatter=2.0.9=py39_torch_1.11.0_cpu
pytorch-sparse=0.6.13=py39_torch_1.11.0_cpu
pytorch-spline-conv=1.2.1=py39_torch_1.11.0_cpu
pytz=2021.3=pypi_0
pywin32=302=py39h2bbff1b_2
pywinpty=2.0.2=py39h5da7b33_0
pyyaml=6.0=py39h2bbff1b_1
pyzmq=22.3.0=py39hd77b12b_2
qt=5.12.9=h5909a2a_4
redis=3.5.3=pypi_0
regex=2022.4.24=pypi_0
requests=2.26.0=pypi_0
ruamel_yaml=0.15.100=py39h2bbff1b_0
s3transfer=0.5.0=pypi_0
scikit-learn=1.0.2=py39hf11a4ad_1
scipy=1.7.3=py39h0a974cb_0
send2trash=1.8.0=pyhd8ed1ab_0
service-identity=21.1.0=pypi_0
setuptools=61.2.0=py39haa95532_0
six=1.16.0=pyhd3eb1b0_1
soupsieve=2.3.1=pyhd8ed1ab_0
sqlite=3.38.3=h2bbff1b_0
sqlparse=0.4.2=pypi_0
stack_data=0.2.0=pyhd3eb1b0_0
terminado=0.13.1=py39haa95532_0
threadpoolctl=2.2.0=pyh0d69192_0
tinycss2=1.1.1=pyhd8ed1ab_0
tk=8.6.12=h8ffe710_0
tomli=2.0.1=pypi_0
torchaudio=0.11.0=py39_cpu
torchvision=0.12.0=py39_cpu
tornado=6.1=py39h2bbff1b_0
tqdm=4.64.0=py39haa95532_0
traitlets=5.1.1=pyhd3eb1b0_0
twisted=21.7.0=pypi_0
twisted-iocpsupport=1.0.2=pypi_0
txaio=21.2.1=pypi_0
typing-extensions=4.0.1=pypi_0
typing_extensions=4.1.1=pyh06a4308_0
tzdata=2021.5=pypi_0
unicodedata2=14.0.0=py39hb82d6ee_1
urllib3=1.26.9=py39haa95532_0
vc=14.2=h21ff451_1
vine=5.0.0=pyhd3eb1b0_0
vs2015_runtime=14.27.29016=h5e58377_2
wcwidth=0.1.8=pypi_0
webencodings=0.5.1=py_1
wheel=0.37.1=pyhd3eb1b0_0
widgetsnbextension=3.6.0=py39hcbf5309_0
win_inet_pton=1.1.0=py39haa95532_0
wincertstore=0.2=py39haa95532_2
winpty=0.4.3=4
wolframclient=1.1.7=pyhd8ed1ab_0
xz=5.2.5=h62dcd97_1
yacs=0.1.6=pyhd3eb1b0_1
yaml=0.2.5=he774522_0
yarl=1.6.3=py39h2bbff1b_0
zipp=3.8.0=py39haa95532_0
zlib=1.2.11=vc14h1cdd9ab_1
zope-interface=5.4.0=pypi_0
zstd=1.4.9=h6255e5f_0
